Move the PureDataPatch class into its own file
