+++ src/pure_data_audio_stream_player.cpp
@@ -51,23 +51,6 @@ static void _print(const char *s)
 	UtilityFunctions::print(s);
 }
 
-void PureDataPatch::_bind_methods()
-{
-	ClassDB::bind_method(D_METHOD("open"), &PureDataPatch::open);
-	ClassDB::bind_method(D_METHOD("close"), &PureDataPatch::close);
-}
-
-bool PureDataPatch::open(String path)
-{
-	handle_ = ::libpd_openfile(path.get_file().utf8().get_data(), path.get_base_dir().utf8().get_data());
-	return handle_ != nullptr;
-}
-
-void PureDataPatch::close()
-{
-	::libpd_closefile(handle_);
-}
-
 void PureDataAudioStream::_bind_methods()
 {
 	ClassDB::bind_method(D_METHOD("is_initialized"), &PureDataAudioStream::is_initialized);
